fix: wfx does not start if built without plugin tag
