+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `wfx` would not start if it was built without plugins support
+
 ### Changed
 
 - Use zstd instead of xz to compress release tarballs
